+++ packages/flutter_login_firebase/lib/src/firebase_login.dart
@@ -3,9 +3,9 @@
 //
 // SPDX-License-Identifier: BSD-3-Clause
 
-import 'package:firebase_auth/firebase_auth.dart';
-import 'package:flutter/widgets.dart';
-import 'package:flutter_login_service/flutter_login_service.dart';
+import "package:firebase_auth/firebase_auth.dart";
+import "package:flutter/widgets.dart";
+import "package:flutter_login_service/flutter_login_service.dart";
 
 /// A service to handle authentication with Firebase.
 /// Takes an instance of [FirebaseAuth] and an
@@ -59,7 +59,7 @@ class FirebaseLoginService implements LoginServiceInterface {
       );
     } on FirebaseAuthException catch (e) {
       loginError = switch (e.code) {
-        'user-disabled' => LoginResponse(
+        "user-disabled" => LoginResponse(
             userObject: null,
             loginSuccessful: false,
             loginError: Error(
@@ -67,7 +67,7 @@ class FirebaseLoginService implements LoginServiceInterface {
               message: translationsService.userDisabledMessage,
             ),
           ),
-        'wrong-password' => LoginResponse(
+        "wrong-password" => LoginResponse(
             userObject: null,
             loginSuccessful: false,
             loginError: Error(
@@ -75,7 +75,7 @@ class FirebaseLoginService implements LoginServiceInterface {
               message: translationsService.wrongPasswordMessage,
             ),
           ),
-        'too-many-requests' => LoginResponse(
+        "too-many-requests" => LoginResponse(
             userObject: null,
             loginSuccessful: false,
             loginError: Error(
@@ -83,12 +83,28 @@ class FirebaseLoginService implements LoginServiceInterface {
               message: translationsService.tooManyRequestsMessage,
             ),
           ),
-        'invalid-credential' => LoginResponse(
+        "invalid-credential" => LoginResponse(
             userObject: null,
             loginSuccessful: false,
             loginError: Error(
               title: translationsService.invalidCredentialsTitle,
               message: translationsService.invalidCredentialsMessage,
+            ),
+          ),
+        "user-not-found" => LoginResponse(
+            userObject: null,
+            loginSuccessful: false,
+            loginError: Error(
+              title: translationsService.invalidCredentialsTitle,
+              message: translationsService.invalidCredentialsMessage,
+            ),
+          ),
+        "network-request-failed" => LoginResponse(
+            userObject: null,
+            loginSuccessful: false,
+            loginError: Error(
+              title: translationsService.networkTitle,
+              message: translationsService.networkMessage,
             ),
           ),
         _ => LoginResponse(
@@ -133,14 +149,14 @@ class FirebaseLoginService implements LoginServiceInterface {
       return const RequestPasswordResponse(requestSuccesfull: true);
     } on FirebaseAuthException catch (e) {
       requestPasswordError = switch (e.code) {
-        'invalid-email' => RequestPasswordResponse(
+        "invalid-email" => RequestPasswordResponse(
             requestSuccesfull: false,
             requestPasswordError: Error(
               title: translationsService.changePasswordInvalidEmailTitle,
               message: translationsService.changePasswordInvalidEmailMessage,
             ),
           ),
-        'user-not-found' => RequestPasswordResponse(
+        "user-not-found" => RequestPasswordResponse(
             requestSuccesfull: false,
             requestPasswordError: Error(
               title: translationsService.changePasswordUserNotFoundTitle,