+++ packages/flutter_login_service/lib/src/login_translations.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import "package:flutter/material.dart";
 
 /// [LoginServiceTranslations] is used to provide custom error messages
 /// for the [LoginService] implementations.
@@ -21,33 +21,37 @@ class LoginServiceTranslations {
     required this.changePasswordUserNotFoundMessage,
     required this.changePasswordErrorTitle,
     required this.changePasswordErrorMessage,
+    required this.networkTitle,
+    required this.networkMessage,
   });
 
   /// [empty] is used to create an empty [LoginServiceTranslations] object, to
   /// use the default error messages.
   const LoginServiceTranslations.empty()
-      : errorTitle = 'Error',
-        errorMessage = 'An unknown error occured',
-        userDisabledTitle = 'User disabled',
+      : errorTitle = "Error",
+        errorMessage = "An unknown error occured",
+        userDisabledTitle = "User disabled",
         userDisabledMessage =
-            'If you have any questions as to how this happened, please contact'
-                ' us at [email].',
-        wrongPasswordTitle = 'Wrong password',
+            "If you have any questions as to how this happened, please contact"
+                " us at [email].",
+        wrongPasswordTitle = "Wrong password",
         wrongPasswordMessage =
-            'The password you entered is incorrect. Please try again.',
-        tooManyRequestsTitle = 'To many requests',
-        tooManyRequestsMessage = 'You’ve tried to log in too many times.',
-        invalidCredentialsTitle = 'Incorrect',
+            "The password you entered is incorrect. Please try again.",
+        tooManyRequestsTitle = "To many requests",
+        tooManyRequestsMessage = "You’ve tried to log in too many times.",
+        invalidCredentialsTitle = "Incorrect",
         invalidCredentialsMessage =
-            'Your login credentials are incorrect, please try again.',
-        changePasswordInvalidEmailTitle = 'Invalid email',
+            "Your login credentials are incorrect, please try again.",
+        changePasswordInvalidEmailTitle = "Invalid email",
         changePasswordInvalidEmailMessage =
-            'The email you entered is invalid. Please try again.',
-        changePasswordUserNotFoundTitle = 'Unknown email',
+            "The email you entered is invalid. Please try again.",
+        changePasswordUserNotFoundTitle = "Unknown email",
         changePasswordUserNotFoundMessage =
-            'The email you entered is unknown. Please try again.',
-        changePasswordErrorTitle = 'Error',
-        changePasswordErrorMessage = 'An unknown error occured';
+            "The email you entered is unknown. Please try again.",
+        changePasswordErrorTitle = "Error",
+        changePasswordErrorMessage = "An unknown error occured",
+        networkTitle = "No network connection",
+        networkMessage = "Please check your network connection";
   final String errorTitle;
   final String errorMessage;
   final String userDisabledTitle;
@@ -64,6 +68,8 @@ class LoginServiceTranslations {
   final String changePasswordUserNotFoundMessage;
   final String changePasswordErrorTitle;
   final String changePasswordErrorMessage;
+  final String networkTitle;
+  final String networkMessage;
 
   /// [copyWith] is used to create a new [LoginServiceTranslations] object
   /// with the provided values. If a value is not provided, the original
@@ -85,6 +91,8 @@ class LoginServiceTranslations {
     String? changePasswordUserNotFoundMessage,
     String? changePasswordErrorTitle,
     String? changePasswordErrorMessage,
+    String? networkTitle,
+    String? networkMessage,
   }) =>
       LoginServiceTranslations(
         errorTitle: errorTitle ?? this.errorTitle,
@@ -108,5 +116,7 @@ class LoginServiceTranslations {
             this.changePasswordUserNotFoundMessage,
         changePasswordErrorTitle: this.changePasswordErrorTitle,
         changePasswordErrorMessage: this.changePasswordErrorMessage,
+        networkTitle: this.networkTitle,
+        networkMessage: this.networkMessage,
       );
 }