+++ packages/flutter_login_firebase/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 5.0.0
+
+* ready the package for deployment to the pub server.
+* Upgraded `flutter_iconica_analysis` to 7.0.0
+* Added new `LoginResponse` object to the `logInWithEmailAndPassword` method.
+
 ## 4.0.0
 
 * Implement isLoggedIn() method from the updated login service to check if the user is logged in.